List item mixin sample
